+++ back/src/bsffs/validation.ts
@@ -152,6 +152,26 @@ export const transporterSchemaFn: FactorySchemaOf<
   Transporter
 > = ({ transporterSignature }) => {
   return yup.object({
+    transporterTransportPlates: yup
+      .array()
+      .of(yup.string())
+      .max(2, "Un maximum de 2 plaques d'immatriculation est accepté")
+      .test((transporterTransportPlates, ctx) => {
+        const { transporterTransportMode } = ctx.parent;
+
+        if (
+          transporterSignature &&
+          transporterTransportMode === "ROAD" &&
+          (!transporterTransportPlates ||
+            !transporterTransportPlates?.filter(p => Boolean(p)).length)
+        ) {
+          return new yup.ValidationError(
+            "La plaque d'immatriculation est requise"
+          );
+        }
+
+        return true;
+      }),
     transporterCompanyName: yup
       .string()
       .requiredIf(