+++ back/src/forms/validation.ts
@@ -787,7 +787,24 @@ export const transporterSchemaFn: FactorySchemaOf<
 > = ({ transporterSignature }) =>
   yup.object({
     transporterCustomInfo: yup.string().nullable(),
-    transporterNumberPlate: yup.string().nullable(),
+    transporterNumberPlate: yup
+      .string()
+      .nullable()
+      .test((transporterNumberPlate, ctx) => {
+        const { transporterTransportMode } = ctx.parent;
+
+        if (
+          transporterSignature &&
+          transporterTransportMode === "ROAD" &&
+          !transporterNumberPlate
+        ) {
+          return new yup.ValidationError(
+            "La plaque d'immatriculation est requise"
+          );
+        }
+
+        return true;
+      }),
     transporterCompanyName: yup
       .string()
       .ensure()