+++ Changelog.md
@@ -5,9 +5,7 @@ Les changements importants de Trackdéchets sont documentés dans ce fichier.
 Le format est basé sur [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 et le projet suit un schéma de versionning inspiré de [Calendar Versioning](https://calver.org/).
 
-
 # [2023.7.1] 18/07/2023
-
 
 #### :rocket: Nouvelles fonctionnalités
 
@@ -16,6 +14,8 @@ et le projet suit un schéma de versionning inspiré de [Calendar Versioning](ht
 - Correction de la date de refus dans les mails de notification et PDFs [PR 2527](https://github.com/MTES-MCT/trackdechets/pull/2527)
 
 #### :boom: Breaking changes
+
+- La plaque d'immatriculation est désormais obligatoire à la signature du transporteur (BSDD, BSFF, BSDASRI, BSDA) [PR 2528](https://github.com/MTES-MCT/trackdechets/pull/2528)
 
 #### :nail_care: Améliorations